fix caching and server actions folder structure
