+++ src/app/(home)/discover/page.tsx
@@ -1,11 +1,7 @@
 import EventList from "@/components/discover/EventList";
 
 function Events() {
-	return (
-		<>
-			<EventList campus="all" />
-		</>
-	);
+	return <EventList campus="all" />;
 }
 
 export default Events;
