+++ src/app/api/user/posts/route.ts
@@ -1,6 +1,7 @@
 import { auth } from "@/auth";
 import prisma from "@/lib/prisma";
 import { checkRateLimit } from "@/lib/server-only";
+import { unstable_cache } from "next/cache";
 import { type NextRequest, NextResponse } from "next/server";
 
 export async function GET(request: NextRequest) {
@@ -14,25 +15,30 @@ export async function GET(request: NextRequest) {
 			return NextResponse.json({ error: "Too many requests" }, { status: 429 });
 		}
 
-		const posts = await prisma.event.findMany({
-			where: {
-				authorId: session.user?.id,
+		const posts = await unstable_cache(
+			async () =>
+				await prisma.event.findMany({
+					where: {
+						authorId: session.user?.id,
+					},
+					select: {
+						id: true,
+						title: true,
+						organizer: true,
+						location: true,
+						poster_url: true,
+						isActive: true,
+					},
+					orderBy: {
+						isActive: "desc",
+					},
+				}),
+			["user-posts", session.user?.id],
+			{
+				revalidate: 60 * 60, // 1 hour
+				tags: ["events", "user-posts"],
 			},
-			select: {
-				id: true,
-				title: true,
-				organizer: true,
-				location: true,
-				poster_url: true,
-				isActive: true,
-			},
-			orderBy: {
-				isActive: "desc",
-			},
-			cacheStrategy: {
-				ttl: 10,
-			},
-		});
+		)();
 
 		return NextResponse.json(posts);
 	} catch (error) {
